requirements.txt
fastapi
uvicorn
pillow
pytesseract
google-generativeai
python-multipart